requirements.txt
lokutor>=1.0.0
python-dotenv
numpy
scipy
pyaudio
websocket-client
websockets
requests
fastapi
uvicorn
twilio
telnyx
